requirements.txt
streamlit
gspread
google-auth
matplotlib
numpy
